Struct local_channel::mpsc::Sender

pub struct Sender<T> { /* private fields */ }

The transmission end of a channel.

This is created by the channel function.

Implementations

impl<T> Sender<T>

pub fn send(&self, item: T) -> Result<(), SendError<T>>

Sends the provided message along this channel.

pub fn close(&mut self)

Closes the sender half.

This prevents any further messages from being sent on the channel, by any sender, while still enabling the receiver to drain messages that are already buffered.